+++ src/Entity.cpp
@@ -36,8 +36,8 @@ Entity::Entity(TombRaider& tr, unsigned int index, unsigned int i, unsigned int 
 }
 
 bool Entity::operator<(Entity& o) {
-    float distA = getRender().mViewVolume.getDistToSphereFromNear(pos[0], pos[1], pos[2], 1.0f);
-    float distB = getRender().mViewVolume.getDistToSphereFromNear(o.pos[0], o.pos[1], o.pos[2], 1.0f);
+    float distA = getRender().getDistToSphereFromNear(pos[0], pos[1], pos[2], 1.0f);
+    float distB = getRender().getDistToSphereFromNear(o.pos[0], o.pos[1], o.pos[2], 1.0f);
     return (distA < distB);
 }